LayoutTests/fast/webgpu/nocrash/fuzz-293854-expected.txt

CONSOLE MESSAGE: promises created
CONSOLE MESSAGE: the end
CONSOLE MESSAGE: fuzz-293854.html
CONSOLE MESSAGE: Pass
